update to suffolk
moved cache/suffolk6578.php to a variable, to minimize errors in
changing, and switched from networksciencelab to suffolk page 6578
(faculty list). had to add u’\u2013’ and u’\u2019’ to the replace list.
If dimitry ever looks at this, help with this problem would be great.
Will be bringing it up in class. Not sure why others didn’t have
problems with this, I got the code from the site…