Expanding code
Counters to implement the addition of names and ranks to the csv file.
will need to expand on this.
Still need to implement opening webpage to get the rest of the
information. shouldn’t be too hard?
use with open(urlX), use soup to find list (or paragraph for the arts
section, will deal with that later. for now, going to skip arts) and
then take each li from the ul and add to the tuple containing name and
rank
putting in this order:
{name, rank, bs, ms, PhD, Institution}
Institution I’m guessing means what department?