Completed
Besides maybe a bit of cleanup in the code, I think I’ve managed to
complete the assignment. Any questions or thoughts on the file let me
know. I’m going to turn this in sometime this afternoon. Also, if you
see any changes you’d like to make just make a branch.